+++ go.mod
@@ -8,7 +8,7 @@ require github.com/rthornton128/goncurses v0.0.0-20220628231859-fd57939296e5
 
 require golang.org/x/exp v0.0.0-20221031165847-c99f073a8326
 
-require gopkg.in/ini.v1 v1.67.0 // indirect
+require gopkg.in/ini.v1 v1.67.0
 
 require github.com/fsnotify/fsnotify v1.6.0
 
@@ -21,18 +21,23 @@ require (
 	github.com/txn2/txeh v1.4.0 // indirect
 	golang.org/x/mod v0.11.0 // indirect
 	golang.org/x/net v0.11.0 // indirect
-	golang.org/x/sys v0.11.0 // indirect
+	golang.org/x/sys v0.16.0 // indirect
 	golang.org/x/tools v0.10.0 // indirect
-	gopkg.in/natefinch/lumberjack.v2 v2.0.0 // indirect
+	gopkg.in/natefinch/lumberjack.v2 v2.0.0
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
-require golang.org/x/term v0.9.0 // indirect
+require (
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	golang.org/x/term v0.9.0 // indirect
+)
 
 require (
 	github.com/clausecker/nfc/v2 v2.1.4
 	github.com/gocarina/gocsv v0.0.0-20230616125104-99d496ca653d
 	github.com/hsanjuan/go-ndef v0.0.1
 	github.com/joho/godotenv v1.4.0
+	github.com/rs/zerolog v1.31.0
 	github.com/wizzomafizzo/mrext v0.0.0-20240102120701-53fb378895e2
 )
