+++ pkg/launcher/commands.go
@@ -18,7 +18,7 @@ You should have received a copy of the GNU General Public License
 along with TapTo.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package tokens
+package launcher
 
 import (
 	"fmt"
