+++ pkg/platforms/mister/config.go
@@ -1,0 +1,12 @@
+package mister
+
+const (
+	TempFolder        = "/tmp/tapto"
+	LogFile           = TempFolder + "/tapto.log"
+	DisableLaunchFile = TempFolder + "/tapto.disabled"
+	SuccessSoundFile  = TempFolder + "/success.wav"
+	FailSoundFile     = TempFolder + "/fail.wav"
+	SocketFile        = TempFolder + "/tapto.sock"
+	MappingsFile      = "/media/fat/nfc.csv"
+	TokenReadFile     = "/tmp/TOKENREAD"
+)
