+++ pkg/tokens/cards.go
@@ -19,7 +19,7 @@ You should have received a copy of the GNU General Public License
 along with TapTo.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package main
+package tokens
 
 import (
 	"encoding/hex"
@@ -35,7 +35,11 @@ const (
 	READ_COMMAND  = byte(0x30)
 )
 
-func getCardUID(target nfc.Target) string {
+var SupportedCardTypes = []nfc.Modulation{
+	{Type: nfc.ISO14443a, BaudRate: nfc.Nbr106},
+}
+
+func GetCardUID(target nfc.Target) string {
 	var uid string
 	switch target.Modulation() {
 	case nfc.Modulation{Type: nfc.ISO14443a, BaudRate: nfc.Nbr106}:
@@ -61,7 +65,7 @@ func comm(pnd nfc.Device, tx []byte, replySize int) ([]byte, error) {
 	return rx, nil
 }
 
-func getCardType(target nfc.Target) string {
+func GetCardType(target nfc.Target) string {
 	switch target.Modulation() {
 	case nfc.Modulation{Type: nfc.ISO14443a, BaudRate: nfc.Nbr106}:
 		var card = target.(*nfc.ISO14443aTarget)
