+++ pkg/tokens/ndef_test.go
@@ -18,7 +18,7 @@ You should have received a copy of the GNU General Public License
 along with TapTo.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package main
+package tokens
 
 import (
 	"bytes"
